Make PLY reading more general. (#24)
Switch PLY reading away from assuming that x, y, z are the same datatype
and consecutive in the file and - if there is color - that it comes
directly after the position. It also opens up the possibility of
understanding and using other less common properties in the PLY files,
like intensity or laser_id. But for now this still does not properly
skip unknown properties.

The new method uses a vector of function pointer that is constructed at
header parsing time. Each function in the vector parses one of the
property and advances the buffer pointer. In my benchmarks this is ~10%
faster than needing to branch for each point.

Tested by diffing a created octree against the same created by master.