Adds a skeleton SDL2 based native viewer with superior performance. (#26)
This is a first draft: It does not update the points it displays yet and the octree directory is hard coded to `./octree`, but it is a working start.

OpenGL is used rather down-to-the-metal on purpose. My hope is that we can get this to compile to emscripten and maybe replace the typescript based one eventually.