Flip Y-Axis of X-Rays to get a right handed coordinate system. (#169)
This means that the images origin is now at the bottom left of the images, i.e. the images also use a right handed coordinate system.

Also a grab bag of unrelated other small refactorings:

Unifying the way the two javascript directories are build.
Add a few missing functions that are in the octree, but not the quadtree crate.
Make the lowest level of the X-Rays use nearest interpolation, so that the actual resolution of the X-Rays is visible.
Bump the X-Ray quadtree version, since the tiles are flipped.