Changed the GRPC client interface to a callback based one to support streaming (#181)
A more ergonomic interface would require async, which is more work that I am willing to put in currently.

Before, all points returned from gRPC where streamed into a Vec, which then was passed on to the consumers. Now the consumers need to provide a callback function that takes a bunch of Points (~100000 per round) and the function can return a bool to indicate if it wants to continue (true) or stop (false) being called with new points.