Fix rotating with the mouse.
It was very slow due to the latest changes for constant rotation speed.

The mouse has a different quality though: you want to follow the mouse
movement as much as possible, i.e. the current rotation speed is
entirely ignored. This renames the delta_angle to rotation speed and
adds a new absolute rotation change request that is only used for mouse
drags.