Initialize bbox with first element (#225)
The bounding box is wrong for point clouds that don't contain (0, 0, 0). This solution may result in a small performance hit. Should I add a TODO to add a next() method to the stream? Then we can get the first point to initialize the bbox from that without adding a check inside the loop.