Pulls out CT mode into open source. (#227)
CT mode sets the near and far plane really close to each other. This allows to look at a slice of the point cloud directly in front of the observer and has been useful for internal debugging at Lyft.

This open sources the implementation, but does not mention the mode in README.md, since it seems very specific.