Support for XBox 360 controller (USB only). (#245)
Only tested on Mac.

You need to install a driver from here before this will work:
https://github.com/360Controller/360Controller/releases

This is not very flexible and it could be that under Linux the scaling works differently so this does not feel nice. However, since this is only used for demos I did not feel like adding dynamic configuration for the scaling values of the different controllers into the code at this point in time.