Oriented beam query (#249)
This is an implementation of oriented beam queries. An oriented beam is the same thing as an ortho frustum, so it is not as specific as it might sound.

Much like the box query, it only opens nodes intersected by the queried volume to be efficient. To do this, a specialized implementation of a collision check between two oriented cubes/boxes is provided.