Xrays with local transforms. (#271)
This PR adds the ability to pass a transform to `build_xray_quadtree`, so all processed points are treated as points in a local axis aligned bounding box. This is important, if the octree points are e.g. in ECEF coordinates, since xray generation expects a local, gravity aligned coordinate frame.