Move data provider out of octree (#314)
The data provider is actually independent of the point cloud representation (as long as it has nodes). This PR strips the "Octree" prefix from DataProvider and moves the module out of octree. The trait interface had to be changed slightly to take node_id as a&str instead of an octree::NodeId, but it was always converted to a String anyway.