Make batch_iterator not directly depend on octree (#315)
BatchIterator is now based on a new PointCloud trait that can be implemented by S2-based pointclouds in the future. Since impl Trait is disallowed in trait definitions, I removed the type parameter from FilteredPointsIterator and specialized it to only filter by PointCulling's contains method.

I moved the batch_iterator module to the top level as well and renamed intersects to intersects_aabb3, to contrast it with the planned intersects_s2_cell method.